refactor(node): extract state subcommands to commands/state.rs
Three cmd_state_* functions move out of main.rs:
  cmd_state_export, cmd_state_import, cmd_state_verify

Same pattern as #641 (wallet) / #643 (validator) / #644 (chain):
pure CLI handlers, no consensus path touched. The heavy comment block
on state_import — documenting why the bail-on-non-zero-height + env
override + trie-reset pattern looks the way it does (2026-04-21 3-way
fork incident, v2.1.5 fix, PR #206 boot integrity check) — stays with
the code; future edits need to see the why.

main.rs drops from 4102 to 3981 lines. cmd_start (BFT engine + libp2p
swarm + validator key load) still inline since the orchestration layer
needs to drive it directly.
